Altered from Ben Tidswell's runLandro2
This version of the code that runs on the Arduino Mega 2560's of Landro L17A has some efficiency edits, and notes on what other aspects of the code need to be verified and finalized.  The review stage is not complete, and may be fleshed out in the form of further comments and edits in future commits.